+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 VSideBar
 
-- `automaticContentSafeAreaEdges(interfaceOrientation:)` method is added in UI model that atomically calculates `contentSafeAreaEdges`
+- `defaultContentSafeAreaEdges(interfaceOrientation:)` method is added in UI model that atomically calculates `contentSafeAreaEdges`
 - Issue with SideBar dismissing without animation on drag-back is fixed
 
 ### [5.0.0(55)](https://github.com/VakhoKontridze/VComponents/releases/tag/5.0.0) — *2023 10 08*